menambhkan permision di command clear

+++ commands/clear.js
@@ -2,11 +2,15 @@ module.exports = {
     name: 'clear',
     description: 'menghapus pesan dalam text channel',
     execute(Message, args) {
-        if (!args[1]) {
-            return Message.reply('masukan jumlah chat yang akan dihapus')
+        var role= Message.member.roles.cache.find(r => r.name ==='🔱Assistant of Hoak🔱')
+        if (role) {
+            if (!args[1]) {
+                return Message.reply('masukan jumlah chat yang akan dihapus')
+            } else {
+            Message.channel.bulkDelete(args[1])
+            }
         } else {
-        Message.channel.bulkDelete(args[1])
+            Message.channel.send('Kamu tidak punya akses ini')
         }
     }
-
 }
